Fixes
Search is fixed.  I had accidentally deleted the @using statement in my
razor code that captures the input.  I also fixed the navbar, literally
and figuratively.  I made my own navbar  and didn't put navbar-fixed-top
in the style portion of the div tag.  It is now present.  It also fixed
my issue of having a lot of blank space at the top.  Apparently, that
space is the space that would have been taken up by the navbar.  Because
now the extra space is gone, and things look good.